requirements.txt
flask
facenet-pytorch
scikit-learn
numpy
pillow
torch
torchvision
werkzeug
